feat: Refactor panels to improve UI structure and integrate find/replace functionality
