+++ infrastructure/entry-points/gui/src/main/java/co/com/leronarenwino/editor/OutputPanel.java
@@ -69,7 +69,7 @@ public class OutputPanel extends EditorPanel {
         bottomPanel.add(positionLabel);
 
         add(new JLabel(labelText), BorderLayout.NORTH);
-        add(scrollPane, BorderLayout.CENTER);
+        add(centerPanel, BorderLayout.CENTER);
         add(bottomPanel, BorderLayout.SOUTH);
     }
