+++ infrastructure/entry-points/gui/src/main/java/co/com/leronarenwino/editor/TemplatePanel.java
@@ -57,7 +57,7 @@ public class TemplatePanel extends EditorPanel {
         bottomPanel.add(positionLabel);
 
         add(new JLabel(labelText), BorderLayout.NORTH);
-        add(scrollPane, BorderLayout.CENTER);
+        add(centerPanel, BorderLayout.CENTER);
         add(bottomPanel, BorderLayout.SOUTH);
     }
